+++ src/application-insights/azext_applicationinsights/commands.py
@@ -14,7 +14,8 @@ from azext_applicationinsights._client_factory import (
     cf_components,
     cf_api_key,
     cf_component_billing,
-    cf_component_linked_storage_accounts
+    cf_component_linked_storage_accounts,
+    cf_export_configuration
 )
 
 
@@ -69,6 +70,16 @@ def load_command_table(self, _):
         client_factory=cf_component_linked_storage_accounts
     )
 
+    export_configurations_sdk = CliCommandType(
+        operations_tmpl='azext_applicationinsights.vendored_sdks.mgmt_applicationinsights.operations.export_configurations_operations#ExportConfigurationsOperations.{}',
+        client_factory=cf_export_configuration
+    )
+
+    export_configurations_custom_sdk = CliCommandType(
+        operations_tmpl='azext_applicationinsights.custom#{}',
+        client_factory=cf_export_configuration
+    )
+
     with self.command_group('monitor app-insights component', command_type=components_sdk, custom_command_type=components_custom_sdk) as g:
         g.custom_command('create', 'create_or_update_component')
         g.custom_command('update', 'update_component')
@@ -101,3 +112,10 @@ def load_command_table(self, _):
         g.custom_command('link', 'create_component_linked_storage_account')
         g.custom_command('update', 'update_component_linked_storage_account')
         g.custom_command('unlink', 'delete_component_linked_storage_account')
+
+    with self.command_group('monitor app-insights component continues-export', command_type=export_configurations_sdk, custom_command_type=export_configurations_custom_sdk) as g:
+        g.custom_command('list', 'list_export_configurations')
+        g.custom_show_command('show', 'get_export_configuration')
+        g.custom_command('create', 'create_export_configuration')
+        g.custom_command('update', 'update_export_configuration')
+        g.custom_command('delete', 'delete_export_configuration', confirmation=True)