+++ src/attestation/azext_attestation/__init__.py
@@ -9,19 +9,22 @@
 # --------------------------------------------------------------------------
 
 from azure.cli.core import AzCommandsLoader
-from azext_attestation.generated._help import helps  # pylint: disable=unused-import
+try:
+    from azext_attestation.manual._help import helps  # pylint: disable=reimported
+except ImportError:
+    pass
 
 
 class AttestationManagementClientCommandsLoader(AzCommandsLoader):
 
     def __init__(self, cli_ctx=None):
         from azure.cli.core.commands import CliCommandType
-        from azext_attestation.generated._client_factory import cf_attestation_mgmt
+        from azext_attestation.generated._client_factory import cf_attestation_cl
         attestation_custom = CliCommandType(
             operations_tmpl='azext_attestation.custom#{}',
-            client_factory=cf_attestation_mgmt)
-        super(AttestationManagementClientCommandsLoader, self).__init__(cli_ctx=cli_ctx,
-                                                                        custom_command_type=attestation_custom)
+            client_factory=cf_attestation_cl)
+        parent = super(AttestationManagementClientCommandsLoader, self)
+        parent.__init__(cli_ctx=cli_ctx, custom_command_type=attestation_custom)
 
     def load_command_table(self, args):
         from azext_attestation.generated.commands import load_command_table