+++ src/connectedk8s/azext_connectedk8s/custom.py
@@ -47,7 +47,7 @@ def create_connectedk8s(cmd, client, resource_group_name, cluster_name, https_pr
     subscription_id = get_subscription_id(cmd.cli_ctx)
 
     # Send cloud information to telemetry
-    send_cloud_telemetry(cmd)
+    azure_cloud = send_cloud_telemetry(cmd)
 
     # Fetching Tenant Id
     graph_client = _graph_client_factory(cmd.cli_ctx)
@@ -72,6 +72,8 @@ def create_connectedk8s(cmd, client, resource_group_name, cluster_name, https_pr
                                 summary='Proxy cert path does not exist')
         raise CLIError(str.format(consts.Proxy_Cert_Path_Does_Not_Exist_Error, proxy_cert))
 
+    proxy_cert = proxy_cert.replace('\\', r'\\\\')
+
     # Checking whether optional extra values file has been provided.
     values_file_provided = False
     values_file = os.getenv('HELMVALUESPATH')
@@ -88,6 +90,7 @@ def create_connectedk8s(cmd, client, resource_group_name, cluster_name, https_pr
     dp_endpoint_dogfood = None
     release_train_dogfood = None
     if cmd.cli_ctx.cloud.endpoints.resource_manager == consts.Dogfood_RMEndpoint:
+        azure_cloud = consts.Azure_DogfoodCloudName
         dp_endpoint_dogfood, release_train_dogfood = validate_env_file_dogfood(values_file, values_file_provided)
 
     # Loading the kubeconfig file in kubernetes client configuration
@@ -192,8 +195,11 @@ def create_connectedk8s(cmd, client, resource_group_name, cluster_name, https_pr
     if os.getenv('HELMREPONAME') and os.getenv('HELMREPOURL'):
         utils.add_helm_repo(kube_config, kube_context)
 
+    # Setting the config dataplane endpoint
+    config_dp_endpoint = get_config_dp_endpoint(cmd, location)
+
     # Retrieving Helm chart OCI Artifact location
-    registry_path = os.getenv('HELMREGISTRY') if os.getenv('HELMREGISTRY') else utils.get_helm_registry(cmd, location, dp_endpoint_dogfood, release_train_dogfood)
+    registry_path = os.getenv('HELMREGISTRY') if os.getenv('HELMREGISTRY') else utils.get_helm_registry(cmd, config_dp_endpoint, dp_endpoint_dogfood, release_train_dogfood)
 
     # Get azure-arc agent version for telemetry
     azure_arc_agent_version = registry_path.split(':')[1]
@@ -231,14 +237,20 @@ def create_connectedk8s(cmd, client, resource_group_name, cluster_name, https_pr
     # Install azure-arc agents
     helm_install_release(chart_path, subscription_id, kubernetes_distro, resource_group_name, cluster_name,
                          location, onboarding_tenant_id, http_proxy, https_proxy, no_proxy, proxy_cert, private_key_pem, kube_config,
-                         kube_context, no_wait, values_file_provided, values_file)
+                         kube_context, no_wait, values_file_provided, values_file, azure_cloud)
 
     return put_cc_response
 
 
 def send_cloud_telemetry(cmd):
-    cloud_name = cmd.cli_ctx.cloud.name
-    telemetry.add_extension_event('connectedk8s', {'Context.Default.AzureCLI.AzureCloud': cloud_name})
+    telemetry.add_extension_event('connectedk8s', {'Context.Default.AzureCLI.AzureCloud': cmd.cli_ctx.cloud.name})
+    cloud_name = cmd.cli_ctx.cloud.name.upper()
+    # Setting cloud name to format that is understood by golang SDK.
+    if cloud_name == consts.PublicCloud_OriginalName:
+        cloud_name = consts.Azure_PublicCloudName
+    elif cloud_name == consts.USGovCloud_OriginalName:
+        cloud_name = consts.Azure_USGovCloudName
+    return cloud_name
 
 
 def validate_env_file_dogfood(values_file, values_file_provided):
@@ -372,6 +384,12 @@ def connected_cluster_exists(client, resource_group_name, cluster_name):
         utils.arm_exception_handler(e, consts.Get_ConnectedCluster_Fault_Type, 'Failed to check if connected cluster resource already exists.', return_if_not_found=True)
         return False
     return True
+
+
+def get_config_dp_endpoint(cmd, location):
+    cloud_based_domain = cmd.cli_ctx.cloud.endpoints.active_directory.split('.')[2]
+    config_dp_endpoint = "https://{}.dp.kubernetesconfiguration.azure.{}".format(location, cloud_based_domain)
+    return config_dp_endpoint
 
 
 def get_public_key(key_pair):
@@ -582,7 +600,7 @@ def get_release_namespace(kube_config, kube_context):
 
 def helm_install_release(chart_path, subscription_id, kubernetes_distro, resource_group_name, cluster_name,
                          location, onboarding_tenant_id, http_proxy, https_proxy, no_proxy, proxy_cert, private_key_pem,
-                         kube_config, kube_context, no_wait, values_file_provided, values_file):
+                         kube_config, kube_context, no_wait, values_file_provided, values_file, cloud_name):
     cmd_helm_install = ["helm", "upgrade", "--install", "azure-arc", chart_path,
                         "--set", "global.subscriptionId={}".format(subscription_id),
                         "--set", "global.kubernetesDistro={}".format(kubernetes_distro),
@@ -590,16 +608,20 @@ def helm_install_release(chart_path, subscription_id, kubernetes_distro, resourc
                         "--set", "global.resourceName={}".format(cluster_name),
                         "--set", "global.location={}".format(location),
                         "--set", "global.tenantId={}".format(onboarding_tenant_id),
-                        "--set", "global.httpsProxy={}".format(https_proxy),
-                        "--set", "global.httpProxy={}".format(http_proxy),
-                        "--set", "global.noProxy={}".format(no_proxy),
                         "--set", "global.onboardingPrivateKey={}".format(private_key_pem),
                         "--set", "systemDefaultValues.spnOnboarding=false",
+                        "--set", "global.azureEnvironment={}".format(cloud_name),
                         "--output", "json"]
     # To set some other helm parameters through file
     if values_file_provided:
         cmd_helm_install.extend(["-f", values_file])
-    if proxy_cert != "":
+    if https_proxy:
+        cmd_helm_install.extend(["--set", "global.httpsProxy={}".format(https_proxy)])
+    if http_proxy:
+        cmd_helm_install.extend(["--set", "global.httpProxy={}".format(http_proxy)])
+    if no_proxy:
+        cmd_helm_install.extend(["--set", "global.noProxy={}".format(no_proxy)])
+    if proxy_cert:
         cmd_helm_install.extend(["--set-file", "global.proxyCert={}".format(proxy_cert)])
     if kube_config:
         cmd_helm_install.extend(["--kubeconfig", kube_config])
@@ -712,6 +734,8 @@ def update_agents(cmd, client, resource_group_name, cluster_name, https_proxy=""
                                 summary='Proxy cert path does not exist')
         raise CLIError(str.format(consts.Proxy_Cert_Path_Does_Not_Exist_Error, proxy_cert))
 
+    proxy_cert = proxy_cert.replace('\\', r'\\\\')
+
     # Checking whether optional extra values file has been provided.
     values_file_provided = False
     values_file = os.getenv('HELMVALUESPATH')
@@ -783,8 +807,11 @@ def update_agents(cmd, client, resource_group_name, cluster_name, https_proxy=""
     if os.getenv('HELMREPONAME') and os.getenv('HELMREPOURL'):
         utils.add_helm_repo(kube_config, kube_context)
 
+    # Setting the config dataplane endpoint
+    config_dp_endpoint = get_config_dp_endpoint(cmd, connected_cluster.location)
+
     # Retrieving Helm chart OCI Artifact location
-    registry_path = os.getenv('HELMREGISTRY') if os.getenv('HELMREGISTRY') else utils.get_helm_registry(cmd, connected_cluster.location, dp_endpoint_dogfood, release_train_dogfood)
+    registry_path = os.getenv('HELMREGISTRY') if os.getenv('HELMREGISTRY') else utils.get_helm_registry(cmd, config_dp_endpoint, dp_endpoint_dogfood, release_train_dogfood)
 
     reg_path_array = registry_path.split(':')
     agent_version = reg_path_array[1]
@@ -801,13 +828,16 @@ def update_agents(cmd, client, resource_group_name, cluster_name, https_proxy=""
 
     cmd_helm_upgrade = ["helm", "upgrade", "azure-arc", chart_path,
                         "--reuse-values",
-                        "--set", "global.httpsProxy={}".format(https_proxy),
-                        "--set", "global.httpProxy={}".format(http_proxy),
-                        "--set", "global.noProxy={}".format(no_proxy),
                         "--wait", "--output", "json"]
     if values_file_provided:
         cmd_helm_upgrade.extend(["-f", values_file])
-    if proxy_cert != "":
+    if https_proxy:
+        cmd_helm_upgrade.extend(["--set", "global.httpsProxy={}".format(https_proxy)])
+    if http_proxy:
+        cmd_helm_upgrade.extend(["--set", "global.httpProxy={}".format(http_proxy)])
+    if no_proxy:
+        cmd_helm_upgrade.extend(["--set", "global.noProxy={}".format(no_proxy)])
+    if proxy_cert:
         cmd_helm_upgrade.extend(["--set-file", "global.proxyCert={}".format(proxy_cert)])
     if kube_config:
         cmd_helm_upgrade.extend(["--kubeconfig", kube_config])