+++ src/connectedk8s/HISTORY.rst
@@ -3,6 +3,11 @@
 Release History
 ===============
 
+0.2.8
+++++++
+* Added checks for proxy
+* Updated config dataplane endpoint to support other clouds
+
 0.2.7
 ++++++
 * Fixed dependency version in setup file